Fix delete project intergration on update with empty dto #88

+++ server/apps/projects/use_cases/project/update.py
@@ -196,5 +196,5 @@ class UseCase(BaseUseCase):
                     project=project,
                     defaults={"token": integration_dto["token"]},
                 )
-            else:
+            elif integration_dto is None:
                 model.objects.filter(project=project).delete()
